refactor: move legacy specific payment gateway files into seperate namespace
